pyproject.toml
[project]
name = "mcp-novel-game-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.2.2",
    "mcp[cli]>=1.6.0",
    "pillow>=11.2.1",
    "pyyaml>=6.0.2",
]
